src: Initial support for the FPGA-IOC
A new top entity is available to target the FPGA-IOC carrier board.
Only the main clocks (GTX&clk_sys) are driven and the lines for
the SFP port.
Most of the reset signals were driven by push buttons in the
MRF version for the other carrier board. We have no push buttons,
so it has been included a one-shot reset after power up to avoid
the criticial warnings from Vivado. However, this should be extended
once the module is used as IP and not as an standalone project.